refactoring(Purchase):component DetailPresentation
